+++ resolwe_bio/tests/unit/test_api.py
@@ -22,7 +22,8 @@ class TestDataViewSetFilters(TestCase):
             slug='test-process',
             version='1.0.0',
             contributor=self.contributor,
-            flow_collection='test-schema',
+            entity_type='test-schema',
+            entity_descriptor_schema='test-schema',
             input_schema=[{'name': 'input_data', 'type': 'data:test:', 'required': False}],
             output_schema=[
                 {'name': 'source', 'type': 'basic:string:'},