+++ docs/CHANGELOG.rst
@@ -44,6 +44,7 @@ Added
 Fixed
 -----
 - Fix Cuffnorm test
+- Fix tests using deprecated flow_collection
 
 
 =================